Update to pass all specs for Crystal 0.31.1
LLVM: 9.0.0
Default target: x86_64-pc-linux-gnu

+++ spec/sysinfo/stat_spec.cr
@@ -1,7 +1,6 @@
 require "../spec_helper"
 
 describe Sysinfo::Stat do
-
   context %<the "cpus" attribute> do
     it "is not empty" do
       Sysinfo::Stat.new.cpus.empty?.should_not be_true
@@ -9,29 +8,28 @@ describe Sysinfo::Stat do
     end
   end
 
-  {% for field in { :intr, :softirq } %}
+  {% for field in {:intr, :softirq} %}
 
     context %<the "{{ field.id }}" attribute> do
       it "is an array of integers" do
         tdata = Sysinfo::Stat.new.{{field.id}}
         tdata.empty?.should_not be_true
-        tdata.should be_a Array(Int32)
+        tdata.should be_a Array(Int64)
         tdata = Sysinfo::Stat.{{field.id}}
         tdata.empty?.should_not be_true
-        tdata.should be_a Array(Int32)
+        tdata.should be_a Array(Int64)
       end
     end
 
   {% end %}
-  {% for field in { :ctxt, :btime, :processes, :procs_running, :procs_blocked } %}
+  {% for field in {:ctxt, :btime, :processes, :procs_running, :procs_blocked} %}
 
   context %<the "{{ field.id }}" attribute> do
     it "is an integer" do
-      Sysinfo::Stat.new.{{field.id}}.should be_a(Int32)
-      Sysinfo::Stat.{{field.id}}.should be_a(Int32)
+      Sysinfo::Stat.new.{{field.id}}.should be_a(Int64)
+      Sysinfo::Stat.{{field.id}}.should be_a(Int64)
     end
   end
 
   {% end %}
-
 end
